Set useInfiniteScroll hook
